requirements.txt
flask==1.1.2
numpy==1.23.2
beautifulsoup4==4.6.3
requests==2.20.1
scipy==1.9.1
pandas==1.4.3
lxml==4.9.1
line-bot-sdk==1.19.0
datetime==4.5
matplotlib==3.5.3
gspread==4.0.1
oauth2client==4.1.3
pyimgur==0.6.0
mpl_finance==0.10.1
arrow==1.2.2